[DOC] help.rst usage.rst README.rst: Reordening some of the base doc info. [CODE] memwiper_core.c memwiper_utils.c: renamed directory [CODE] All python files import requalified an ordered.
