+++ src/memwiper/__init__.py
@@ -1,5 +1,6 @@
-from memwiper_core import wipeit
-import memwiper_utils as utils
+import memwiper.c_src.memwiper_utils as utils
+from memwiper.c_src.memwiper_core import wipeit
+
 __all__ = [
     "wipeit",
     "utils",
