+++ src/memwiper/utils/__init__.py
@@ -1,4 +1,6 @@
-from memwiper_utils import kind, codepoints, size
+from memwiper.c_src.memwiper_utils import codepoints
+from memwiper.c_src.memwiper_utils import kind
+from memwiper.c_src.memwiper_utils import size
 
 __all__ = [
     "kind",
