+++ tests/test_memwiper.py
@@ -1,4 +1,5 @@
 import pytest
+
 import memwiper
